backend/requirements.txt
# NeuroCapture Backend Dependencies
# Python requirements for the FastAPI backend service

# Core Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.20.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.28.0
alembic>=1.10.0

# Data Validation and Processing
pydantic>=2.0.0
python-dotenv>=1.0.0

# Audio Processing and Analysis
librosa>=0.10.0
noisereduce>=3.0.0
praat-parselmouth>=0.4.0
webrtcvad>=2.0.0
soundfile>=0.12.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
matplotlib>=3.7.0

# Utilities and Tools
tqdm>=4.65.0
httpx>=0.24.0

# Testing Framework
pytest>=7.0.0
pytest-asyncio>=0.21.0
aiosqlite>=0.19.0  # For testing with SQLite
